Adding SetFrom to IEmailConfigurator
