+++ README.md
@@ -8,7 +8,7 @@
 
 > The second letter in the Hebrew alphabet is the ב bet/beit. Its meaning is "house". In the ancient pictographic Hebrew it was a symbol resembling a tent on a landscape.
 
-_Note: Pre-release packages are distributed via [feedz.io](https://f.feedz.io/kdcllc/bet-extensions/nuget/index.json)._
+_Note: Pre-release packages are distributed via [feedz.io](https://f.feedz.io/kdcllc/bet-notifications/nuget/index.json)._
 
 This goal of this repo is to provide with a reusable functionality for developing Microservices with Docker and Kubernetes.
 
@@ -22,6 +22,8 @@ Please send [email](mailto:kingdavidconsulting@gmail.com) if you consider to **h
 
 If you like or are using this project to learn or start your solution, please give it a star. Thanks!
 
+
+## Usage
 
 
 ## Resources
