--- src/Bet.Notifications.Razor/Bet - Backup.Notifications.Razor.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-		<TargetFrameworks>netstandard2.0;net6.0</TargetFrameworks>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Bet.Notifications\Bet.Notifications.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.0.0" />
-    <PackageReference Include="RazorLight" Version="2.0.0-rc.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Repository\EntityFrameworkCore\" />
-  </ItemGroup>
-
-</Project>
